# Collapse revision context into specialized context to avoid allocations