<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

Always wanted to play with custom value contexts, so here we are! This spares us a bit of linked list lookup and context allocation, so yay!

```
benchmark                                                         old ns/op     new ns/op     delta
BenchmarkContextHandler/context_handler_success-sequential-16     923           841           -8.88%
BenchmarkContextHandler/context_handler_success-parallel-16       176           166           -5.68%
BenchmarkContextHandler/context_handler_failure-sequential-16     15401         11938         -22.49%
BenchmarkContextHandler/context_handler_failure-parallel-16       11216         11829         +5.47%

benchmark                                                         old allocs     new allocs     delta
BenchmarkContextHandler/context_handler_success-sequential-16     7              6              -14.29%
BenchmarkContextHandler/context_handler_success-parallel-16       7              6              -14.29%
BenchmarkContextHandler/context_handler_failure-sequential-16     38             38             +0.00%
BenchmarkContextHandler/context_handler_failure-parallel-16       38             38             +0.00%

benchmark                                                         old bytes     new bytes     delta
BenchmarkContextHandler/context_handler_success-sequential-16     592           560           -5.41%
BenchmarkContextHandler/context_handler_success-parallel-16       592           560           -5.41%
BenchmarkContextHandler/context_handler_failure-sequential-16     3944          4111          +4.23%
BenchmarkContextHandler/context_handler_failure-parallel-16       3663          3942          +7.62%
```

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```

/assign @julz @vagababov